add unit tests for the student model
